Fix snapshotting + d3 bugs
For some reason we couldn't just do:
state.nodes = snapshot.nodes
Instead we had to loop through the snapshot nodes and call the addNode
function with the in-game defined state. This had the side effect of
updating seenNodes, ensuring we've updated the user node appropriately
(since the person doing the snapshot is not the person reading the
node). The same holds for edges/seenEdges.

No idea why this fixed it, but it did!